check for http response error in ExtractInto*Ptr methods
rename tenantattr pkg to volumetenants
